New lib dashboard export + move cell renderer utilities from cmponents export
