+++ packages/admin/dashboard/src/exports/components.tsx
@@ -43,20 +43,6 @@ export const LayoutComposer = Object.assign(LayoutComposerRoot, {
 export type { LayoutComponentProps } from "../components/layout-composer/types"
 
 export {
-  defineCellRenderer,
-  type CellRenderer,
-  type CellRendererDefinition,
-  type CellAlignment,
-  type RenderMode,
-  type BuiltInRenderMode,
-} from "../lib/table/cell-renderers"
-
-export {
   ConfigurableDataTable,
   type ConfigurableDataTableProps,
 } from "../components/table/configurable-data-table/configurable-data-table"
-
-export {
-  createTableAdapter,
-  type TableAdapter,
-} from "../lib/table/table-adapters"
