+++ packages/admin/dashboard/src/exports/lib.ts
@@ -1,0 +1,13 @@
+export {
+  defineCellRenderer,
+  type CellRenderer,
+  type CellRendererDefinition,
+  type CellAlignment,
+  type RenderMode,
+  type BuiltInRenderMode,
+} from "../lib/table/cell-renderers"
+
+export {
+  createTableAdapter,
+  type TableAdapter,
+} from "../lib/table/table-adapters"
