+++ packages/admin/dashboard/tsup.config.cjs
@@ -8,6 +8,8 @@ export default defineConfig({
     components: "./src/exports/components.tsx",
     // `@medusajs/dashboard/hooks`
     hooks: "./src/exports/hooks.ts",
+    // `@medusajs/dashboard/lib`
+    lib: "./src/exports/lib.ts",
   },
   format: ["cjs", "esm"],
   external: [
@@ -27,6 +29,7 @@ export default defineConfig({
       index: "./src/index.ts",
       components: "./src/exports/components.tsx",
       hooks: "./src/exports/hooks.ts",
+      lib: "./src/exports/lib.ts",
     },
   },
   clean: true,
